2025-08-05 changed(after chat FIX)

+++ src/main/java/com/springmvc/controller/VoteController.java
@@ -108,8 +108,14 @@ public class VoteController {
 
     @PostMapping("/delete")
     public String deleteVote(@RequestParam("voteId") int voteId, HttpSession session) {
+        System.out.println("투표 삭제 컨트롤러 " + voteId);
         Member mb = (Member) session.getAttribute("mb");
-        voteService.deleteVoteById(voteId);
-        return "redirect:/voteList";
+        if (mb != null) {
+            voteService.deleteVoteById(voteId);
+        } else {
+            return "redirect:member/login";
+        }
+
+        return "redirect:/vote/list";
     }
 }
